Sync modules, added conditional cleaning of files

+++ Makefile
@@ -1,5 +1,5 @@
 CC = gcc
-C_FLAGS =-Wall -Wextra  -fsanitize=address -g
+C_FLAGS = -Wall -Wextra -g
 
 dps_src_path := $(or $(DPS_ROOT), .)
 dps_slave.c := $(dps_src_path)/dps_slave.c
@@ -9,21 +9,32 @@ ifndef $(C_VECTOR_ROOT)
 C_VECTOR_ROOT := $(dps_src_path)/lib/c_vector
 include $(C_VECTOR_ROOT)/Makefile
 endif
+all: release
 
-all: master slave
+DEBUG_FLAGS = -O0 -fsanitize=address
+RELEASE_FLAGS = -O2
 
-master: dps_master.o
-slave : dps_slave.o
+debug: C_FLAGS += $(DEBUG_FLAGS)
+debug: compile
 
-dps_master.o: $(dps_master.c)
-	$(CC) $(C_FLAGS) $(dps_master.c) -c
+release: C_FLAGS += $(RELEASE_FLAGS)
+release: compile
 
-dps_slave.o: $(dps_slave.c)
-	$(CC) $(C_FLAGS) $(dps_slave.c) -c
+
+compile: dps_master.o dps_slave.o
+
+dps_master.o: $(dps_master.c) 
+	$(CC) $(C_FLAGS) $(DEBUG) $(dps_master.c) -c
+
+dps_slave.o: $(dps_slave.c) 
+	$(CC) $(C_FLAGS) $(DEBUG) $(dps_slave.c) -c
 
 dps_clean:
+ifeq ($(wildcard dps_slave.o), dps_slave.o)
 	rm dps_slave.o
+endif
+ifeq ($(wildcard dps_master.o), dps_master.o)
 	rm dps_master.o
-
+endif
 
 clean: dps_clean
